+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## 4.3.0 - Unreleased
+### Added
+- [backport] Added `maxTotalMergeKeys` (10000) loader option to limit the total number of
+  keys processed by YAML merge (`<<`) across one `load()` / `loadAll()` call.
+
+### Removed
+- [backport] `maxMergeSeqLength` replaced with `maxTotalMergeKeys` for limiting YAML merge
+  processing.
+
+
 ## [4.2.0] - 2026-06-01
 ### Added
 - Added `docs/safety.md` with notes about processing untrusted YAML.